IGNITE-22410 review fixes
